Create plugin as a filter which allows analyze difference between two time steps of one data array (field).
Added the following server manager properties:
1. Property is responsible for enumeration the scalar array names of input data set;
2. Property is responsible for specifying the first time step;
3. Property is responsible for specifying the second time step;
4. Property is responsible for specifying the prefix of new array name.

Filter is capable of reading various data objects.
Also filter can return the corresponding errors if the input data are not correct or property of user interface controls is invalid. This filter appointed into "Temporal" group of filters.
Added the icon to this filter.